Refactor IndexSerde
